Reorganize test cases
